+++ src/components/MultiSeriesBarChart/utilities/tests/get-min-max.test.tsx
@@ -1,5 +1,6 @@
 import {Series, StackSeries} from 'components/MultiSeriesBarChart/types';
 
+import {DEFAULT_MAX_Y} from '../../../../constants';
 import {getMinMax} from '../get-min-max';
 
 const mockData: Series[] = [
@@ -35,6 +36,39 @@ const mockStackedData = [
   ],
 ] as StackSeries[];
 
+const mockZeroData: Series[] = [
+  {
+    data: [
+      {label: 'label', rawValue: 0},
+      {label: 'label', rawValue: 0},
+      {label: 'label', rawValue: 0},
+    ],
+    color: 'colorBlack',
+    name: 'LABEL1',
+  },
+  {
+    data: [
+      {label: 'label', rawValue: 0},
+      {label: 'label', rawValue: 0},
+      {label: 'label', rawValue: 0},
+    ],
+    color: 'colorBlack',
+    name: 'LABEL2',
+  },
+];
+const mockZeroStackedData = [
+  [
+    [0, 0],
+    [0, 0],
+    [0, 0],
+  ],
+  [
+    [0, 0],
+    [0, 0],
+    [0, 0],
+  ],
+] as StackSeries[];
+
 describe('get-min-max', () => {
   it('returns min and max of non stacked data when stackedValues is null', () => {
     const {min, max} = getMinMax(null, mockData);
@@ -48,5 +82,15 @@ describe('get-min-max', () => {
 
     expect(min).toStrictEqual(0);
     expect(max).toStrictEqual(33);
+  });
+
+  it('returns the default max y value for non stacked values of all zeros', () => {
+    const {max} = getMinMax(null, mockZeroData);
+    expect(max).toStrictEqual(DEFAULT_MAX_Y);
+  });
+
+  it('returns the default max y value for stacked values of all zeros', () => {
+    const {max} = getMinMax(mockZeroStackedData, mockZeroData);
+    expect(max).toStrictEqual(DEFAULT_MAX_Y);
   });
 });